+++ CHANGES.md
@@ -1,7 +1,15 @@
 2.1.0 (in development)
 ======================
 
+- Support for pickling `abc.abstractproperty`, `abc.abstractclassmethod`,
+  and `abc.abstractstaticmethod`.
+  ([PR #450](https://github.com/cloudpipe/cloudpickle/pull/450))
 
+- Support for pickling subclasses of generic classes.
+  ([PR #448](https://github.com/cloudpipe/cloudpickle/pull/448))
+
+- Support and CI configuration for Python 3.11.
+  ([PR #467](https://github.com/cloudpipe/cloudpickle/pull/467))
 
 2.0.0
 =====